Clean tests and code
