+++ .coveragerc
@@ -1,2 +1,10 @@
 [run]
 branch = True
+
+[report]
+exclude_lines =
+    pragma: no cover
+    def __repr__
+    raise NotImplementedError
+omit =
+    .*
