+++ .gitignore
@@ -1,10 +1,10 @@
 *.pyc
 *.egg-info
+/.cache
 /.coverage
-/htmlcov
+/.eggs
 /build
 /dist
-/.tox
-/docs/_build
 /env
+/htmlcov
 /profile
