+++ cssselect2/compiler.py
@@ -10,7 +10,6 @@ from webencodings import ascii_lower
 from . import parser
 from .parser import SelectorError
 
-
 # http://dev.w3.org/csswg/selectors/#whitespace
 split_whitespace = re.compile('[^ \t\r\n\f]+').findall
 
@@ -20,8 +19,8 @@ def compile_selector_list(input, namespaces=None):
 
     :param input:
         A :term:`tinycss2:string`,
-        or an iterable of tinycss2 :term:`tinycss2:component values`
-        such as the :attr:`~tinycss2.ast.QualifiedRule.predule` of a style rule.
+        or an iterable of tinycss2 :term:`tinycss2:component values` such as
+        the :attr:`~tinycss2.ast.QualifiedRule.predule` of a style rule.
     :param namespaces:
         A optional dictionary of all `namespace prefix declarations
         <http://www.w3.org/TR/selectors/#nsdecl>`_ in scope for this selector.
@@ -208,7 +207,7 @@ def _compile_node(selector):
     elif isinstance(selector, parser.PseudoClassSelector):
         if selector.name == 'link':
             return ('%s and el.get_attr("href") is not None'
-                     % html_tag_eq('a', 'area', 'link'))
+                    % html_tag_eq('a', 'area', 'link'))
         elif selector.name == 'enabled':
             return (
                 '(%s and el.get_attr("disabled") is None'
@@ -234,9 +233,10 @@ def _compile_node(selector):
             )
         elif selector.name == 'checked':
             return (
-                '(%s and el.get_attr("checked") is not None and '
-                'ascii_lower(el.get_attr("type", "")) in ("checkbox", "radio"))'
-                ' or (%s and el.get_attr("selected") is not None)'
+                '(%s and el.get_attr("checked") is not None and'
+                ' ascii_lower(el.get_attr("type", "")) '
+                ' in ("checkbox", "radio"))'
+                'or (%s and el.get_attr("selected") is not None)'
                 % (
                     html_tag_eq('input', 'menuitem'),
                     html_tag_eq('option'),
