+++ cssselect2/tests/__init__.py
@@ -1,0 +1,11 @@
+# coding: utf8
+"""
+    cssselect2.tests
+    ----------------
+
+    Test suite for cssselect2.
+
+    :copyright: (c) 2012 by Simon Sapin.
+    :license: BSD, see LICENSE for more details.
+
+"""
