+++ cssselect2/tests/ids.html
@@ -1,0 +1,48 @@
+<html id="html" xmlns="http://www.w3.org/1999/xhtml"><head>
+    <link id="link-href" href="foo" />
+    <link id="link-nohref" />
+  </head><body>
+    <div id="outer-div">
+      <a id="name-anchor" name="foo"></a>
+      <a id="tag-anchor" rel="tag" href="http://localhost/foo">link</a>
+      <a id="nofollow-anchor" rel="nofollow" href="https://example.org">
+        link</a>
+      <ol id="first-ol" class="a b c">
+        <li id="first-li">content</li>
+        <li id="second-li" lang="En-us">
+          <div id="li-div">
+          </div>
+        </li>
+        <li id="third-li" class="ab c"></li>
+        <li id="fourth-li" class="ab
+c"></li>
+        <li id="fifth-li"></li>
+        <li id="sixth-li"></li>
+        <li id="seventh-li">  </li>
+      </ol>
+      <p id="paragraph">
+        <b id="p-b">hi</b> <em id="p-em">there</em>
+        <b id="p-b2">guy</b>
+        <input type="checkbox" id="checkbox-unchecked" />
+        <input type="checkbox" id="checkbox-disabled" disabled="" />
+        <input type="text" id="text-checked" checked="checked" />
+        <input type="hidden" id="input-hidden" />
+        <input type="hidden" id="input-hidden-disabled" disabled="disabled" />
+        <input type="checkbox" id="checkbox-checked" checked="checked" />
+        <input type="checkbox" id="checkbox-disabled-checked"
+               disabled="disabled" checked="checked" />
+        <fieldset id="fieldset" disabled="disabled">
+          <input type="checkbox" id="checkbox-fieldset-disabled" />
+          <input type="hidden" id="hidden-fieldset-disabled" />
+        </fieldset>
+      </p>
+      <ol id="second-ol">
+      </ol>
+      <map name="dummymap">
+        <area shape="circle" coords="200,250,25" href="foo.html" id="area-href" />
+        <area shape="default" id="area-nohref" />
+      </map>
+    </div>
+    <div id="foobar-div" foobar="ab bc
+cde"><span id="foobar-span"></span></div>
+</body></html>
