+++ setup.cfg
@@ -1,10 +1,6 @@
-[build_sphinx]
-source-dir = docs
-build-dir  = docs/_build
-#all_files  = 1
+[aliases]
+test = pytest
 
-[upload_sphinx] # Sphinx-PyPI-upload
-upload-dir = docs/_build/html
-
-[pytest]
-python_files = tests.py
+[tool:pytest]
+addopts = --cov --flake8 --isort cssselect2
+norecursedirs = dist .cache .git build *.egg-info .eggs venv
