+++ setup.py
@@ -1,15 +1,18 @@
 # coding: utf8
 
-import re
 import os.path
-from setuptools import setup
+import re
+import sys
 
+from setuptools import setup
 
 ROOT = os.path.dirname(__file__)
 README = open(os.path.join(ROOT, 'README.rst')).read()
 INIT_PY = open(os.path.join(ROOT, 'cssselect2', '__init__.py')).read()
 VERSION = re.search("VERSION = '([^']+)'", INIT_PY).group(1)
 
+needs_pytest = {'pytest', 'test', 'ptr'}.intersection(sys.argv)
+pytest_runner = ['pytest-runner'] if needs_pytest else []
 
 setup(
     name='cssselect2',
@@ -23,4 +26,10 @@ setup(
     packages=['cssselect2'],
     package_data={'cssselect2': ['tests/*']},
     install_requires=['tinycss2'],
+    setup_requires=pytest_runner,
+    test_suite='cssselect2.tests',
+    tests_require=[
+        'pytest-runner', 'pytest-cov', 'pytest-flake8', 'pytest-isort'],
+    extras_require={'test': [
+        'pytest-runner', 'pytest-cov', 'pytest-flake8', 'pytest-isort']},
 )
