--- tox.ini
@@ -1,7 +1,0 @@
-[tox]
-envlist = py26,py27,py32,py33,pypy
-
-[testenv]
-deps = pytest
-changedir = {toxworkdir}/log
-commands = py.test --pyargs cssselect2 []
